+++ carp_context_package/lib/context.dart
@@ -11,11 +11,11 @@ import 'dart:io' show Platform;
 import 'package:json_annotation/json_annotation.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:activity_recognition_alt/activity_recognition_alt.dart';
-import 'package:location/location.dart';
 import 'package:weather/weather.dart';
 import 'package:openmhealth_schemas/openmhealth_schemas.dart' as omh;
 import 'package:permission_handler/permission_handler.dart';
 import 'package:air_quality/air_quality.dart';
+import 'package:geolocator/geolocator.dart';
 
 part 'activity_datum.dart';
 part 'activity_probe.dart';