+++ carp_context_package/lib/location_datum.dart
@@ -15,22 +15,22 @@ class LocationDatum extends CARPDatum {
 
   LocationDatum() : super();
 
-  LocationDatum.fromLocationData(LocationData location)
-      : latitude = location.latitude,
-        longitude = location.longitude,
-        altitude = location.altitude,
-        accuracy = location.accuracy,
-        speed = location.speed,
-        speedAccuracy = location.speedAccuracy,
-        heading = location.heading,
-        time = location.time,
+  LocationDatum.fromPositionData(Position position)
+      : latitude = position.latitude,
+        longitude = position.longitude,
+        altitude = position.altitude,
+        accuracy = position.accuracy,
+        speed = position.speed,
+        speedAccuracy = position.speedAccuracy,
+        heading = position.heading,
+        time = position.timestamp,
         super();
 
   factory LocationDatum.fromJson(Map<String, dynamic> json) => _$LocationDatumFromJson(json);
   Map<String, dynamic> toJson() => _$LocationDatumToJson(this);
 
-  /// The time in milliseconds
-  double time;
+  /// The time when this location was collected.
+  DateTime time;
 
   /// Latitude in GPS coordinates.
   double latitude;
@@ -60,8 +60,7 @@ class LocationDatum extends CARPDatum {
 
   String toString() =>
       super.toString() +
-      ','
-          ' latitude: $latitude, '
+      'latitude: $latitude, '
           'longitude: $longitude, '
           'accuracy; $accuracy, '
           'altitude: $altitude, '